+++ R/get_schema.R
@@ -59,8 +59,19 @@ get_schema.Id <- function(obj,  ...) {
 #' @export
 #' @rdname get_schema
 get_schema.PqConnection <- function(obj, temporary = FALSE,  ...) {
-  if (isTRUE(temporary))  {
-    return(DBI::dbGetQuery(obj, "SELECT nspname FROM pg_namespace WHERE oid = pg_my_temp_schema();")$nspname)
+  if (temporary)  {
+    temp_schema <- DBI::dbGetQuery(obj, "SELECT nspname FROM pg_namespace WHERE oid = pg_my_temp_schema();")$nspname
+
+    # If no temporary tables have been created, the temp schema has not been determined yet and we create a temporary
+    # table to force the schema to be determined.
+    if (identical(temp_schema, character(0))) {
+      temp_table_name <- unique_table_name()
+      DBI::dbExecute(obj, glue::glue("CREATE TEMPORARY TABLE {temp_table_name}(a TEXT)"))
+      temp_schema <- DBI::dbGetQuery(obj, "SELECT nspname FROM pg_namespace WHERE oid = pg_my_temp_schema();")$nspname
+      DBI::dbExecute(obj, glue::glue("DROP TABLE {temp_table_name}"))
+    }
+
+    return(temp_schema)
   } else {
     return(DBI::dbGetQuery(obj, "SELECT CURRENT_SCHEMA()")$current_schema)
   }
@@ -69,7 +80,7 @@ get_schema.PqConnection <- function(obj, temporary = FALSE,  ...) {
 #' @export
 #' @rdname get_schema
 get_schema.SQLiteConnection <- function(obj, temporary = FALSE,  ...) {
-  if (isTRUE(temporary))  {
+  if (temporary)  {
     return("temp")
   } else {
     return("main")